Fix compiler warnings
- Add override to BitmapSurface virtual method overrides
- Add virtual destructor to MR_PolygonShape for safe base-pointer deletion
- Cast enum to double in Missile.cpp float/enum comparison
- Add default cases to Sprite.cpp switches for unhandled eLeft/eTop

+++ ObjFac1/Missile.cpp
@@ -306,7 +306,7 @@ void MR_Missile::ApplyEffect( const MR_ContactEffect* pEffect,  MR_SimulationTim
    if( (lPhysCollision != NULL)&&pValidDirection )
    {
 
-      if( lPhysCollision->mWeight < MR_PhysicalCollision::eInfiniteWeight )
+      if( lPhysCollision->mWeight < static_cast<double>(MR_PhysicalCollision::eInfiniteWeight) )
       {
          if( mLived >= cIgnitionTime )
          {
